requirements.txt
PyMuPDF==1.23.7
pytesseract==0.3.10
Pillow==10.0.0
